5. Longest Palindromic Substring
Solution:
idea - We pass on every letter in the string and check if it is the center of a palindrome.
we have two cases, odd length palindrome and even length. we use two indices L and R to compare the two sides of the palindrome.
Case 1: The letter is a center of odd-length palindrome, initialize R and L to be indices to that letter, and start comparing then increase r and decrease L until the letters that L and R point to are not equal. 
update the max length and the palindrome accordingly.
Case 2: the letter is a center of an even-length palindrome, same code but the initialization for R and L are R=L+1=i+1, and the rest of the checking is the same.