17. Letter Combinations of a Phone Number
Given a string containing digits from 2-9 inclusive, return all possible letter combinations that the number could represent. Return the answer in any order.

A mapping of digits to letters (just like on the telephone buttons) is given below. Note that 1 does not map to any letters.

Solution idea: 
we can easily associate a string of letters to each digit using a switch case.
now assume we have the vector of strings we want the combination of, we use a recursive function to get these combinations.
we look at the first string in the vector, assume that the function already gives us the combination of the rest of the strings in the vector, pass on every letter in the first string append the combination to it, and return the result.
classic approach to building a recursive function makes the solution easier to implement.