49. Group Anagrams
Q: Given an array of strings strs, group the anagrams together. You can return the answer in any order.

An Anagram is a word or phrase formed by rearranging the letters of a different word or phrase, typically using all the original letters exactly once.

Solution:
the idea behind my solution, pass on every str using a histogram to count the letters, then pass on the rest of the words and check if the histogram never goes below 0, if so add it to the result of anagrams of the current str.
if we use a used array that keeps track of which strs we already added, we will get rid of redundant checks on future strs that already were added as anagrams to previous strs.
The downside of this solution is it passes on the histogram twice,
so for every word we do 26 extra steps for counting the numbers, and 26 more for every other word to check if it's an anagram. which is 26 squared. so the complexity of the solution is O(26*26*N*M) where N is the number of strings, and M is the max string size.
there is another solution that involves sorting each word instead of using a histogram, and with the help of the map we can get the anagrams, the complexity of such a solution would take O(NLOG(M)), so if we know that M<9 then its better sorting than using a histogram.
note: 9 because logM < 26*26=676   ---> M < 9.